code cleanup

+++ classes/task/tool_inactive_user_cleanup_task.php
@@ -21,7 +21,6 @@
  * @author DualCube <admin@dualcube.com>
  * @copyright DualCube (https://dualcube.com)
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
- * tool_inactive_user_cleanup is standard cron function
  */
 namespace tool_inactive_user_cleanup\task;
 
@@ -48,47 +47,40 @@ class tool_inactive_user_cleanup_task extends \core\task\scheduled_task {
         mtrace(get_string('taskstart' , 'tool_inactive_user_cleanup'));
         $beforedelete = get_config('tool_inactive_user_cleanup' , 'daysbeforedeletion');
         $inactivity = get_config('tool_inactive_user_cleanup', 'daysofinactivity');
-        if ($inactivity > 0) {
-            $subject = get_config('tool_inactive_user_cleanup', 'emailsubject');
-            $body = get_config('tool_inactive_user_cleanup', 'emailbody');
-            $users = $DB->get_records('user', ['deleted' => '0']);
-            $messagetext = html_to_text($body);
-            $mainadminuser = get_admin();
-            foreach ($users as $usersdetails) {
-                $minus = round((time() - $usersdetails->lastaccess) / 60 / 60 / 24);
-                if ($minus > $inactivity) {
-                    $ischeck = $DB->get_record('tool_inactive_user_cleanup', ['userid' => $usersdetails->id]);
-                    if (!$ischeck &&  $usersdetails->lastaccess != 0) {
-                        $record = new \stdClass();
-                        $record->userid = $usersdetails->id;
-                        if (email_to_user($usersdetails, $mainadminuser, $subject, $messagetext)) {
-                            mtrace(get_string('userid' , 'tool_inactive_user_cleanup'));
-                            mtrace($usersdetails->id. '---' .$usersdetails->email);
-                            mtrace(get_string('userinactivtime' , 'tool_inactive_user_cleanup') . $minus);
-                            mtrace('');
-                            $record->emailsent = 1;
-                            $record->date = time();
-                            $DB->insert_record('tool_inactive_user_cleanup' , $record, false);
-                        }
-                    }
-                }
-                if ($beforedelete != 0 &&  $usersdetails->lastaccess != 0) {
-                    $deleteuserafternotify = $DB->get_record('tool_inactive_user_cleanup' , ['userid' => $usersdetails->id]);
-                    if (!empty($deleteuserafternotify)) {
-                        $mailssent = $deleteuserafternotify->date;
-                        $diff = round((time() - $mailssent) / 60 / 60 / 24);
-                        if ($diff > $beforedelete) {
-                            if (!isguestuser($usersdetails->id)) {
-                                delete_user($usersdetails);
-                                mtrace(get_string('deleteduser' , 'tool_inactive_user_cleanup') . $usersdetails->id);
-                                mtrace(get_string('detetsuccess' , 'tool_inactive_user_cleanup'));
-                            }
-                        }
-                    }
+        if ($inactivity == 0) {
+            mtrace(get_string('invalaliddayofinactivity' , 'tool_inactive_user_cleanup'));
+            return;
+        }
+        $subject = get_config('tool_inactive_user_cleanup', 'emailsubject');
+        $body = get_config('tool_inactive_user_cleanup', 'emailbody');
+        $users = $DB->get_records('user', ['deleted' => '0']);
+        $messagetext = html_to_text($body);
+        $mainadminuser = get_admin();
+        foreach ($users as $usersdetails) {
+            $minus = round((time() - $usersdetails->lastaccess) / 60 / 60 / 24);
+            $ischeck = $DB->get_record('tool_inactive_user_cleanup', ['userid' => $usersdetails->id]);
+            $record = new \stdClass();
+            $record->userid = $usersdetails->id;
+            if ($minus > $inactivity && !$ischeck && $usersdetails->lastaccess != 0 && email_to_user($usersdetails, $mainadminuser, $subject, $messagetext)) {
+                mtrace(get_string('userid' , 'tool_inactive_user_cleanup'));
+                mtrace($usersdetails->id. '---' .$usersdetails->email);
+                mtrace(get_string('userinactivtime' , 'tool_inactive_user_cleanup') . $minus);
+                mtrace('');
+                $record->emailsent = 1;
+                $record->date = time();
+                $DB->insert_record('tool_inactive_user_cleanup' , $record, false);
+            }
+            if ($beforedelete != 0 &&  $usersdetails->lastaccess != 0) {
+                $deleteuserafternotify = $DB->get_record('tool_inactive_user_cleanup' , ['userid' => $usersdetails->id]);
+                $beforedelete = get_config('tool_inactive_user_cleanup' , 'daysbeforedeletion');
+                $mailssent = $deleteuserafternotify->date;
+                $diff = round((time() - $mailssent) / 60 / 60 / 24);
+                if (!empty($deleteuserafternotify) && $diff > $beforedelete && !isguestuser($usersdetails->id)) {
+                    delete_user($usersdetails);
+                    mtrace(get_string('deleteduser' , 'tool_inactive_user_cleanup') . $usersdetails->id);
+                    mtrace(get_string('detetsuccess' , 'tool_inactive_user_cleanup'));
                 }
             }
-        } else {
-            mtrace(get_string('invalaliddayofinactivity' , 'tool_inactive_user_cleanup'));
         }
         mtrace(get_string('taskend' , 'tool_inactive_user_cleanup'));
     }//end of function execute()
